remove liquidation component - add modal to payment in appoiments
