+++ src/app/appointments/services/appointment.service.ts
@@ -1,14 +1,16 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environment/environment';
-import { Observable } from 'rxjs';
-import { AppointmentDto, TurnoDetalleDTO } from '../models/appointment.model';
+import { Observable, tap } from 'rxjs';
+import { AppointmentDto, TurnoDetalleDTO, Factura, PlanObraSocial, ObraSocial, MetodoPago } from '../models/appointment.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AppointmentService {
   private baseUrl = `${environment.apiUrl}/api/Turno`;
+  private URI = `${environment.apiUrl}/api/`;
+
 
   constructor(private http: HttpClient) {}
 
@@ -38,5 +40,39 @@ export class AppointmentService {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   deleteAppointment(id: number): Observable<any> {
     return this.http.delete(`${this.baseUrl}/AnularTurno/${id}`);
+  }
+
+  
+  registrarFactura(billing: Factura): Observable<Factura> {
+    const body = { ...billing }; 
+    return this.http.post<Factura>(`${this.URI}Factura/RegistrarFactura`, body).pipe(
+      tap(response => { 
+        return response;
+      })
+    );
+  }
+
+  getObrasSociales(): Observable<ObraSocial[]> {
+    return this.http.get<ObraSocial[]>(`${this.URI}ObraSocial/obtenerTodasLasObrasSociales`).pipe(
+      tap(response => { 
+        return response;
+      })
+    );
+  }
+
+  getPlanesObraSocial(idObraSocial: number): Observable<PlanObraSocial[]> {
+    return this.http.get<PlanObraSocial[]>(`${this.URI}PlanObraSocial/obtenerPlanesPorObraSocial/${idObraSocial}`).pipe(
+      tap(response => { 
+        return response;
+      })
+    );
+  }
+
+  getMetodosPago(): Observable<MetodoPago[]> {
+    return this.http.get<MetodoPago[]>(`${this.URI}MetodoDePago/obtenerMetodosDePago`).pipe(
+      tap(response => { 
+        return response;
+      })
+    );
   }
 }
