+++ src/app/billing/components/billing-list/billing-list.component.html
@@ -4,7 +4,7 @@
 <div class="flex flex-col 2xl:flex-row 2xl:gap-x-10 gap-4 mb-5 overflow-hidden">
   <app-custom-table
     class="md:mt-4 2xl:mt-0 flex-1"
-    [title]="'Ultimas facturas'"
+    [title]="'Ultimas Liquidaciones'"
     [data]="tableData"
     [columns]="tableHeaders"
     (onDetailsClick)="handleDetailsClick($event)" />
@@ -72,6 +72,112 @@
           currentLiquidation.numeroTransaccion
         }}</span>
       </p>
+    </div>
+  </div>
+</div>
+<div class="flex flex-col 2xl:flex-row 2xl:gap-x-10 gap-4 mb-5 overflow-hidden">
+  <app-custom-table
+    class="md:mt-4 2xl:mt-0 flex-1"
+    [title]="'Ultimas Facturas'"
+    [data]="tableDataLiq"
+    [columns]="tableHeadersLiq"
+    (onDetailsClick)="handleDetailsClickLiq($event)" />
+
+  <!-- Factura viewer -->
+  <div
+    *ngIf="currentBilling"
+    class="flex flex-col md:mt-4 2xl:mt-0 bg-gray-100 rounded-md p-4 min-w-40">
+    <h6
+      class="block font-sans text-xl font-semibold leading-relaxed text-blue-gray-900 mb-4 text-center">
+      Factura
+    </h6>
+
+    <!-- Información General -->
+    <div class="mb-2">
+      <div class="flex justify-between items-center mb-2">
+        <p class="text-sm text-gray-600">
+          Factura Nº:
+          <span class="font-semibold">{{ currentBilling.idTurno }}</span>
+        </p>
+        <p class="text-sm text-gray-600">
+          Fecha:
+          <span class="font-semibold">{{ currentBilling.fechaPago }}</span>
+        </p>
+      </div>
+      <p class="text-sm text-gray-600">
+        Cliente:
+        <span class="font-semibold">{{ currentBilling.nombrePaciente }}</span>
+      </p>
+      <div class="flex justify-between items-center">
+        <p class="text-sm text-gray-600">
+          Médico:
+          <span class="font-semibold">{{ currentBilling.nombreMedico }}</span>
+        </p>
+        <p class="text-sm text-gray-600">
+          Especialidad:
+          <span class="font-semibold">{{
+            currentBilling.especialidadMedico
+          }}</span>
+        </p>
+      </div>
+    </div>
+    <!-- Información del Turno -->
+    <div class="mt-2">
+      <p class="text-sm text-gray-600">
+        Fecha y Hora del Turno:
+        <span class="font-semibold"
+          >{{ currentBilling.fechaTurno.slice(0, 10) }}
+          {{ currentBilling.horaInicio }} - {{ currentBilling.horaFin }}</span
+        >
+      </p>
+      <p class="text-sm text-gray-600">
+        Motivo: <span class="font-semibold">{{ currentBilling.motivo }}</span>
+      </p>
+      <p class="text-sm text-gray-600">
+        Notas:
+        <span class="font-semibold">{{ currentBilling.notasTurno }}</span>
+      </p>
+    </div>
+
+    <!-- Información de Pago -->
+    <div class="mt-2">
+      <p class="text-sm text-gray-600">
+        Método de Pago:
+        <span class="font-semibold">{{ currentBilling.nombreMetodoPago }}</span>
+      </p>
+      <p class="text-sm text-gray-600">
+        Número de Transacción:
+        <span class="font-semibold">{{
+          currentBilling.numeroTransaccion
+        }}</span>
+      </p>
+      <p class="text-sm text-gray-600">
+        Obra Social:
+        <span class="font-semibold">{{
+          currentBilling.nombreObraSocial +
+            ' - ' +
+            currentBilling.nombrePlanObraSocial
+        }}</span>
+      </p>
+      <p class="text-sm text-gray-600">
+        Cobertura:
+        <span class="font-semibold">{{ currentBilling.cobertura }}%</span>
+      </p>
+    </div>
+
+    <!-- Detalles de Pago -->
+    <div class="flex justify-between items-center bg-white p-4 rounded-md mb-2">
+      <span class="text-sm font-semibold text-gray-700">Total Paciente:</span>
+      <span class="text-sm font-semibold text-gray-900">{{
+        currentBilling.montoPaciente
+      }}</span>
+    </div>
+
+    <div class="flex justify-between items-center bg-white p-4 rounded-md mb-2">
+      <span class="text-sm font-semibold text-gray-700">Total:</span>
+      <span class="text-lg font-bold text-gray-900">{{
+        currentBilling.montoTotal
+      }}</span>
     </div>
   </div>
 </div>
