+++ src/app/billing/models/billing.model.ts
@@ -45,3 +45,43 @@ export interface Medico {
   fechaNacimiento: string;
   horarioDisponible: object;
 }
+
+export interface FacturaDetallada {
+  idFactura: number;
+  ObraSocial: string;
+  numeroTransaccion: string;
+  montoTotal: number;
+  fechaPago: string;
+}
+
+export interface FacturaDetalle {
+  apellidoMedico: string;
+  apellidoPaciente: string;
+  cobertura: number;
+  dniPaciente: number;
+  especialidadMedico: string;
+  estado: string;
+  fechaHorario: string;          
+  fechaPago: string;             
+  fechaTurno: string;            
+  horaFin: string;
+  horaInicio: string;
+  idFactura: number;
+  idMedico: number;
+  idMetodoPago: number;
+  idObraSocial: number;
+  idPaciente: number;
+  idPlanObraSocial: number;
+  idTurno: number;
+  montoPaciente: number;
+  montoTotal: number;
+  motivo: string;
+  nombreMedico: string;
+  nombreMetodoPago: string;
+  nombreObraSocial: string;
+  nombrePaciente: string;
+  nombrePlanObraSocial: string;
+  notasTurno: string;
+  numeroTransaccion: string;
+  precioTurno: number;
+}
