--- src/app/liquidation/components/liquidation-details/liquidation-details.component.html
@@ -1,1 +1,0 @@
-<p>liquidation-details works!</p>
