--- src/app/liquidation/components/liquidation-details/liquidation-details.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-liquidation-details',
-  standalone: true,
-  imports: [],
-  templateUrl: './liquidation-details.component.html',
-  styleUrl: './liquidation-details.component.css'
-})
-export class LiquidationDetailsComponent {
-
-}
