+++ src/app/shared/components/custom-table/custom-table.component.ts
@@ -55,6 +55,7 @@ import { FormsModule } from '@angular/forms';
             </svg>
           </span>
         </button>
+        
       </div>
       <!-- Filtros -->
       <div class="p-4 flex flex-wrap gap-3 items-end">
@@ -118,7 +119,8 @@ import { FormsModule } from '@angular/forms';
                 </p>
               </td>
               <!-- Ícono de más detalles -->
-              <td class="py-3 px-5 border-b border-blue-gray-50 text-center">
+              <td class="py-3 px-5 border-b border-blue-gray-50 text-center flex">
+
                 <button
                   (click)="handleDetailsClick(row)"
                   class="text-blue-500 hover:text-blue-700">
@@ -135,7 +137,27 @@ import { FormsModule } from '@angular/forms';
                       d="M15 12H9m4-4H9m6 8H9" />
                   </svg>
                 </button>
+
+                <button
+                *ngIf="handlePaymentClick"
+                (click)="handlePaymentClick.emit(row)"
+                  class="text-green-500 hover:text-green-800">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke-width="2"
+                    stroke="currentColor"
+                    class="w-5 h-5">
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      d="M15 12H9m4-4H9m6 8H9" />
+                  </svg>
+                </button>
+
               </td>
+
             </tr>
           </tbody>
         </table>
@@ -158,6 +180,7 @@ export class CustomTableComponent implements OnInit {
   // Función personalizada para el ícono de detalles
   // eslint-disable-next-line @angular-eslint/no-output-on-prefix
   @Output() onDetailsClick = new EventEmitter<TableRow>();
+  @Output() handlePaymentClick = new EventEmitter<TableRow>();
 
   filteredData: TableRow[] = [];
   searchTerm = ''; // Nueva propiedad para el texto de búsqueda
